backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic>=2.0.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv